+++ src/modules/cart/components/Cart/Cart.jsx
@@ -1,10 +1,5 @@
-import { useSelector, useDispatch } from 'react-redux';
+import { useSelector } from 'react-redux';
 import { selectProducts, selectTotalPrice } from '@redux/cart/cartSlice';
-import {
-  addOrUpdateProduct,
-  deleteProduct,
-  decreaseQuantity,
-} from '@redux/cart/cartSlice';
 import CartContainer from '../CartContainer/CartContainer';
 import EmptyCart from '../EmptyCartContainer/EmptyCartContainer';
 import sprite from '../../../../shared/icons/sprite.svg';
@@ -13,22 +8,9 @@ import s from './Cart.module.scss';
 const Cart = ({ onClose }) => {
   const products = useSelector(selectProducts);
   const totalPrice = useSelector(selectTotalPrice);
-  const dispatch = useDispatch();
 
   const handleClose = () => {
     onClose();
-  };
-
-  const handleAddOrUpdate = (id) => {
-    dispatch(addOrUpdateProduct({ id: id, quantity: 1 }));
-  };
-
-  const handleDecrease = (id) => {
-    dispatch(decreaseQuantity(id));
-  };
-
-  const handleDelete = (id) => {
-    dispatch(deleteProduct(id));
   };
 
   const handleBackBtn = () => {
@@ -46,9 +28,6 @@ const Cart = ({ onClose }) => {
       {products.length > 0 ? (
         <CartContainer
           products={products}
-          handleIncrease={handleAddOrUpdate}
-          handleDecrease={handleDecrease}
-          handleDelete={handleDelete}
           handleBackBtn={handleBackBtn}
           handleClose={handleClose}
           totalPrice={totalPrice}
